Refactor import paths to use backend module structure; remove setup.sh
